backend-ml/requirements.txt
torch
transformers
Pillow
flask 
werkzeug
flask-cors
google-genai
python-dotenv
pymysql
flask-jwt-extended
bcrypt

